Anjay 1.13.0
Breaking changes:
- anjay_configuration_t::max_icmp_failures field has been removed.
- Changed the way connection errors are handled. Connections are now NOT
  automatically retried in most of the cases. Please refer to the documentation
  (Advanced tutorial -> Network error handling) for a summary of the new
  semantics.

Improvements:
- Extensive refactor of the server connection handling subsystem.
- Added timeout for the documentation URL check.
- Prevented integration tests from running concurrently on Travis.

Bugfixes:
- Fixed behaviour when the attributes are set so that pmax < pmin.
- Fixed a bug that caused Discover operation on the Security object to
  erroneously work instead of causing 4.01 Unauthorized error as mandated by the
  spec.
- Fixed compilation warnings on various compilers.
- Upgraded avs_commons to version 3.8.2, which includes:
  - Fixes for proper propagation of avs_stream_close() errors.
  - Fixes for external library dependency checking.
  - Fixes for various compilation warnings.
  - Improved logs from the IP address stringification code.

+++ src/coap/stream/server_internal.c
@@ -658,9 +658,10 @@ int _anjay_coap_server_read(coap_server_t *server,
 static int
 block_write(coap_server_t *server, const void *data, size_t data_length) {
     if (!server->block_ctx) {
-        uint16_t block_size = server->curr_block.valid
-                                      ? server->curr_block.size
-                                      : AVS_COAP_MSG_BLOCK_MAX_SIZE;
+        uint16_t block_size =
+                (uint16_t) (server->curr_block.valid
+                                    ? server->curr_block.size
+                                    : AVS_COAP_MSG_BLOCK_MAX_SIZE);
 
         server->static_id_source = _anjay_coap_id_source_new_static(
                 _anjay_coap_server_get_request_identity(server));
